renamed blue-eyes references to blue_eyes
