+++ CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# CONTRIBUTING
+
+Bug reports and pull requests are welcome on GitHub at https://github.com/scarver2/blue-eyes. Feel free to contribute, but please consider that any contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct. In summary, be excellent to each other, and party on dudes.
