+++ LICENSE
@@ -1,6 +1,8 @@
-Copyright (c) 2016 Brian Emory
-Copyright (c) 2024 Ben Lamb
-Copyright (c) 2025 Stan Carver II
+License: MIT
+
+Copyright (c) 2016 Brian Emory (Corneal Ruby Gem)
+Copyright (c) 2024 Ben Lamb (blue-eyes Ruby Gem)
+Copyright (c) 2025 Stan Carver II (BlueEyes Ruby Gem)
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
