+++ bin/blue_eyes
