--- lib/blue-eyes.rb
@@ -1,11 +1,0 @@
-require "blue-eyes/extensions/string"
-require "blue-eyes/generators/app/app_generator"
-require "blue-eyes/generators/controller/controller_generator"
-require "blue-eyes/generators/model/model_generator"
-require "blue-eyes/generators/scaffold/scaffold_generator"
-require "blue-eyes/cli"
-require "blue-eyes/version"
-
-module BlueEyes
-
-end
