+++ lib/blue_eyes.rb
@@ -1,0 +1,11 @@
+require "blue_eyes/extensions/string"
+require "blue_eyes/generators/app/app_generator"
+require "blue_eyes/generators/controller/controller_generator"
+require "blue_eyes/generators/model/model_generator"
+require "blue_eyes/generators/scaffold/scaffold_generator"
+require "blue_eyes/cli"
+require "blue_eyes/version"
+
+module BlueEyes
+
+end
