+++ lib/blue_eyes/extensions/string.rb
