+++ lib/blue_eyes/generators/app/templates/README.md
