+++ lib/blue_eyes/generators/app/templates/Rakefile
