+++ lib/blue_eyes/generators/app/templates/config/initializers/database.rb
