+++ lib/blue_eyes/generators/app/templates/config/initializers/redis.rb
