+++ lib/blue_eyes/generators/app/templates/public/favicon.ico
