+++ lib/blue_eyes/generators/app/templates/public/stylesheets/main.css
