+++ lib/blue_eyes/generators/app/templates/spec/spec_helper.rb
