+++ lib/blue_eyes/generators/controller/templates/controller.rb.erb
