+++ lib/blue_eyes/generators/controller/templates/views/show.html.erb
