+++ test/blue_eyes/cli_test.rb
