+++ test/blue_eyes/generators/scaffold/scaffold_generator_test.rb
